+++ appinfo/routes.php
@@ -1,5 +1,6 @@
 <?php
 
+declare(strict_types=1);
 /**
  * Breeze Dark theme for Nextcloud
  *
@@ -25,6 +26,7 @@
  */
 
 return ['routes' => [
+	['name' => 'page#index', 'url' => '/client-details', 'verb' => 'GET', 'postfix' => 'direct.new'],
 	['name' => 'page#index', 'url' => '/', 'verb' => 'GET'],
 	['name' => 'page#create', 'url' => '/create', 'verb' => 'POST'],
 	['name' => 'page#update', 'url' => '/update', 'verb' => 'POST'],
