+++ package.json
@@ -39,7 +39,8 @@
 		"sanitize-html": "^2.7.0",
 		"vue": "^2.6.14",
 		"vue-avatar": "^2.3.3",
-		"vue-linkify": "^1.0.1"
+		"vue-linkify": "^1.0.1",
+		"vue-router": "^3.6.4"
 	},
 	"browserslist": [
 		"extends @nextcloud/browserslist-config"
